src/main/java/com/howard/leetcode/linkedlist/MergeTwoOrderedListsSolution.java
package com.howard.leetcode.linkedlist;

/**
 * 合并两个有序链表
 *
 * @author howard he
 * @create 2018-12-25 05:29
 */
public class MergeTwoOrderedListsSolution {

    /**
     *
     * @param l1
     * @param l2
     * @return
     */
    public ListNode mergeTwoLists(ListNode l1, ListNode l2) {
        ListNode p1 = l1;
        ListNode p2 = l2;
        ListNode dummy = new ListNode(-1);
        ListNode curr = dummy;
        while (p1 != null && p2 != null) {
            if (p1.val <= p2.val) {
                curr.next = p1;
                p1 = p1.next;
            } else {
                curr.next = p2;
                p2 = p2.next;
            }
            curr = curr.next;
        }
        if (p1 != null) {
            curr.next = p1;
        }
        if (p2 != null) {
            curr.next = p2;
        }
        return dummy.next;
    }
}
